Maybe discovers Heisenbug in CUnit?
Running until CU_ASSERT_EQUAL_FATAL(Failures, CU_get_number_of_failures())
always leads to a final result output of 1 failed assert--that very assert--
paradoxical, no?

Because for that assert to be the failure, it would mean that the number of failures
had changed before it was called... correct? Meaning there would need to be at least
2 failed asserts at the end, but always only 1.

When I printed the values being checked for equality in the double loop, they always
printed equal pairs to stderr--so they couldn't be triggering it.

For sanity, I removed the CU_ASSERT_EQUAL_FATAL, and instead printed the number of errors
at the beginning and end of the test, both print 0, and then stack smashing is detected,
crashing the program.

What is going on here?