Refines random_unique_indexes_array
Tested, and working, but permutation inequality bias unsolved--

For example, with indexes 2 and variation 1, there are only 3 possible outcomes:

0 1, 0 2, and 1 2, or additively,
0 0, 0 1, and 1 0.

If all permutations should be equally likely, then we should see the function return the last index as 1 approximately 1/3 of the time. However, manual testing has demonstrated this result consistently 1/4 of the time. It seems unlikely that at such a small scale, the error would come from the pseudo RNG provided by libc in srand and rand. As such, further investigation is required to resolve this issue.