Fixes memory access error in radix_sort by changing 'e' from 'char' to 'unsigned char'
Interestingly, the previous commit was made on a machine running on an RK3399 ARMv8 (aarch64) chip, and all tests pass locally with and without the modification in this commit.
So far, the error has only occurred on x86(_64) machines.

Investigation into the assembly codes generated between the 2 target architectures in conjunction with GCC and ISO C 99-11 specs may yield useful insights into this curious discrepancy